Fix broken tests
