+++ response_operations_ui/templates/messages.html
@@ -7,7 +7,7 @@
 <div role="main" id="main" class="page__main">
   {% if change_survey %}
     <h1 class="jupiter" name="page-messages-title">{{ selected_survey }} Messages
-      <span class="mars">(<a id="change-survey" href="/messages/select-survey">Change survey</a>)</span>
+      <span class="mars">(<a id="change-survey" href="/messages">Change survey</a>)</span>
     </h1>
 
   {% else %}
