+++ tests/views/test_message.py
@@ -7,21 +7,26 @@ from response_operations_ui import app
 from response_operations_ui.controllers.message_controllers import _get_url, send_message
 from response_operations_ui.exceptions.exceptions import InternalError
 
+shortname_url = f'{app.config["BACKSTAGE_API_URL"]}/v1/survey/shortname'
 url_get_thread = f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/threads/fb0e79bd-e132-4f4f-a7fd-5e8c6b41b9af'
+url_sign_in_data = f'{app.config["BACKSTAGE_API_URL"]}/v2/sign-in/'
+url_get_threads_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/threads'
+url_get_surveys_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/survey/surveys'
+
 with open('tests/test_data/message/thread.json') as json_data:
     thread_json = json.load(json_data)
-url_sign_in_data = f'{app.config["BACKSTAGE_API_URL"]}/v2/sign-in/'
 
 with open('tests/test_data/message/thread_missing_subject.json') as json_data:
     thread_missing_subject = json.load(json_data)
-url_get_threads_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/threads'
-url_get_surveys_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/survey/surveys'
 
 with open('tests/test_data/message/threads.json') as json_data:
     thread_list = json.load(json_data)
 
 with open('tests/test_data/survey/survey_list.json') as json_data:
     survey_list = json.load(json_data)
+
+with open('tests/test_data/survey/ashe_response.json') as json_data:
+    ashe_info = json.load(json_data)
 
 
 class TestMessage(unittest.TestCase):
@@ -50,8 +55,9 @@ class TestMessage(unittest.TestCase):
     def test_threads_list(self, mock_request):
         mock_request.get(url_get_threads_list, json=thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Apple".encode(), response.data)
         self.assertIn("50012345678".encode(), response.data)
@@ -64,12 +70,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_atmsg_to(self, mock_request):
         with open('tests/test_data/message/threads_missing_atmsg_to.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -78,12 +83,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_atmsg_from(self, mock_request):
         with open('tests/test_data/message/threads_missing_atmsg_from.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -92,12 +96,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_msg_to(self, mock_request):
         with open('tests/test_data/message/threads_missing_msg_to.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -106,12 +109,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_date(self, mock_request):
         with open('tests/test_data/message/threads_missing_sent_date.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -120,12 +122,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_ru_ref(self, mock_request):
         with open('tests/test_data/message/threads_missing_ru_ref.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -134,12 +135,11 @@ class TestMessage(unittest.TestCase):
     def test_threads_list_with_missing_business_name(self, mock_request):
         with open('tests/test_data/message/threads_missing_business_name.json') as thread_json:
             malformed_thread_list = json.load(thread_json)
-        mock_request.get(url_get_threads_list,
-                         json=malformed_thread_list
-                         )
+        mock_request.get(url_get_threads_list, json=malformed_thread_list)
         mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Unavailable".encode(), response.data)
         self.assertIn("Example message subject".encode(), response.data)
@@ -147,8 +147,9 @@ class TestMessage(unittest.TestCase):
     @requests_mock.mock()
     def test_threads_list_fail(self, mock_request):
         mock_request.get(url_get_threads_list, status_code=500)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages", follow_redirects=True)
+        response = self.app.get("/messages/ASHE", follow_redirects=True)
 
         self.assertEqual(response.status_code, 500)
         self.assertIn("Error 500 - Server error".encode(), response.data)
@@ -156,8 +157,9 @@ class TestMessage(unittest.TestCase):
     @requests_mock.mock()
     def test_threads_list_empty(self, mock_request):
         mock_request.get(url_get_threads_list, json={"messages": []})
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
-        response = self.app.get("/messages")
+        response = self.app.get("/messages/ASHE")
 
         self.assertEqual(response.status_code, 200)
         self.assertIn("No new messages".encode(), response.data)
@@ -190,7 +192,8 @@ class TestMessage(unittest.TestCase):
     def test_request_response_malformed(self, mock_request):
         url = url_get_threads_list
         mock_request.get(url, json={})
-        response = self.app.get("/messages")
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
+        response = self.app.get("/messages/ASHE")
 
         self.assertEqual(response.status_code, 200)
         self.assertIn("Something went wrong".encode(), response.data)
@@ -229,12 +232,15 @@ class TestMessage(unittest.TestCase):
         self.assertIn("Please enter a message".encode(), response.data)
 
     message_form = {'body': "TEST BODY",
-                    'subject': "TEST SUBJECT"}
+                    'subject': "TEST SUBJECT",
+                    'hidden_survey': "ASHE"}
 
     @requests_mock.mock()
     def test_form_submit_with_valid_data(self, mock_request):
         mock_request.post(f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/send-message', status_code=201)
-        mock_request.get(url_get_threads_list, json={}, status_code=200)
+        mock_request.get(url_get_threads_list, json=thread_list, status_code=200)
+        mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+        mock_request.get(shortname_url + "/ASHE", json=ashe_info)
 
         with app.app_context():
             response = self.app.post("/messages/create-message", data=self.message_form, follow_redirects=True)
@@ -249,9 +255,7 @@ class TestMessage(unittest.TestCase):
         with app.app_context():
             response = self.app.post("/messages/create-message", data=self.message_form, follow_redirects=True)
 
-        self.assertIn(
-            "Message failed to send, something has gone wrong with the website.".encode(),
-            response.data)
+        self.assertIn("Message failed to send, something has gone wrong with the website.".encode(), response.data)
         self.assertIn("TEST SUBJECT".encode(), response.data)
         self.assertIn("TEST BODY".encode(), response.data)
 
@@ -307,3 +311,32 @@ class TestMessage(unittest.TestCase):
         response = self.app.get("/messages/threads/fb0e79bd-e132-4f4f-a7fd-5e8c6b41b9af")
         self.assertEqual(response.status_code, 200)
         self.assertIn("No Subject".encode(), response.data)
+
+    def test_get_radio_buttons(self):
+        response = self.app.get("/messages")
+
+        self.assertEqual(200, response.status_code)
+        self.assertIn("Filter messages by survey".encode(), response.data)
+        self.assertIn("ASHE".encode(), response.data)
+        self.assertIn("Bricks".encode(), response.data)
+
+    @requests_mock.mock()
+    def test_radio_buttons_post_nothing_selected(self, mock_request):
+        mock_request.get(url_get_threads_list, json=thread_list)
+        mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
+
+        response = self.app.post("/messages", follow_redirects=True)
+
+        self.assertEqual(200, response.status_code)
+        self.assertIn("Home".encode(), response.data)
+        self.assertIn("Please select a survey to filter your messages.".encode(), response.data)
+        self.assertIn("Filter messages by survey".encode(), response.data)
+
+    @requests_mock.mock()
+    def test_get_messages_survey_does_not_exist(self, mock_request):
+        mock_request.get(shortname_url)
+
+        response = self.app.get("/messages/this-survey", follow_redirects=True)
+
+        self.assertEqual(response.status_code, 500)
+        self.assertIn("Error 500 - Server error".encode(), response.data)
